添加注释

+++ src/main/java/io/github/libkodi/security/entity/Access.java
@@ -4,11 +4,11 @@ import lombok.Data;
 
 @Data
 public class Access {
-	private int count = 0;
+	private long count = 0;
 	private boolean blocked = false;
 	private String ip;
 	
-	public int increment() {
+	public long increment() {
 		count += 1;
 		return count;
 	}
